Voice_Agent/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
sqlalchemy==2.0.22
psycopg2-binary==2.9.9
redis==5.0.1
pydantic==2.4.2
python-dotenv==1.0.0
pydantic-settings==2.8.1
httpx==0.25.0